+++ market-ui/src/Constants.ts
@@ -1,12 +1,9 @@
-export const MICROCCD_IN_CCD = 1000000;
 export const MARKET_CONTRACT_ADDRESS = {
 	index: BigInt(984),
 	subindex: BigInt(0),
 };
 export const MARKET_CONTRACT_SCHEMA =
 	"FFFF02010000000A0000004D61726B65742D4E465400030000000300000061646400140003000000140000006E66745F636F6E74726163745F616464726573730C08000000746F6B656E5F69641D000500000070726963650A040000006C69737401140101000000100114000300000008000000746F6B656E5F69641D0008000000636F6E74726163740C0500000070726963650A080000007472616E7366657200140003000000140000006E66745F636F6E74726163745F616464726573730C08000000746F6B656E5F69641D0002000000746F0B";
-export const MARKET_CONTRACT_SCHEMA_BASE64 =
-	"//8CAQAAAAoAAABNYXJrZXQtTkZUAAMAAAADAAAAYWRkABQAAwAAABQAAABuZnRfY29udHJhY3RfYWRkcmVzcwwIAAAAdG9rZW5faWQdAAUAAABwcmljZQoEAAAAbGlzdAEUAQEAAAAQARQAAwAAAAgAAAB0b2tlbl9pZB0ACAAAAGNvbnRyYWN0DAUAAABwcmljZQoIAAAAdHJhbnNmZXIAFAADAAAAFAAAAG5mdF9jb250cmFjdF9hZGRyZXNzDAgAAAB0b2tlbl9pZB0AAgAAAHRvCw==";
 export const NFT_CONTRACT_MODULE_REF =
 	"8e8e7beaa6a1ced4041077c641efb10bbf0f6b8c4f3614d1861d06a6fc6b40dc";
 export const NFT_CONTRACT_SCHEMA =
