+++ market-ui/src/models/Cis2Client.ts
@@ -5,12 +5,10 @@ import {
 	TransactionSummary,
 	AccountAddress,
 	ModuleReference,
-	GtuAmount,
 } from "@concordium/web-sdk";
 
 import { Cis2Deserializer } from "./Cis2Deserializer";
 import { MetadataUrl, OperatorOfQueryParams } from "./Cis2Types";
-import { parseContractAddress, toGtu } from "./Utils";
 import { NFT_CONTRACT_MODULE_REF, NFT_CONTRACT_SCHEMA } from "../Constants";
 import {
 	initContract,
@@ -30,12 +28,19 @@ export const enum MethodName {
 const SCHEMA_BUFFER = Buffer.from(NFT_CONTRACT_SCHEMA, "hex");
 const MODULE_REF = new ModuleReference(NFT_CONTRACT_MODULE_REF);
 
+/**
+ * Initilizes CIS2-NFT contract.
+ * @param provider Wallet Provider.
+ * @param account Account to initialize the contract with.
+ * @param maxContractExecutionEnergy Max energy to be used to initialize the contract.
+ * @returns
+ */
 export async function initCis2NftContract(
 	provider: WalletApi,
 	account: string,
 	maxContractExecutionEnergy = BigInt(9999)
-) {
-	const outcomes = await initContract(
+): Promise<ContractAddress> {
+	return await initContract(
 		provider,
 		MODULE_REF,
 		SCHEMA_BUFFER,
@@ -43,10 +48,16 @@ export async function initCis2NftContract(
 		account,
 		maxContractExecutionEnergy
 	);
-
-	return parseContractAddress(outcomes);
 }
 
+/**
+ * Checks wether the input market address in an Operator Of the input account in the CIS2 contract.
+ * @param provider Wallet Provider.
+ * @param account Account address.
+ * @param marketAddress Marketplace Contract Address.
+ * @param nftAddress CIS2-NFT contract address.
+ * @returns true if @see marketAddress is an operator of @see account in @see nftAddress
+ */
 export async function isOperator(
 	provider: WalletApi,
 	account: string,
@@ -77,10 +88,16 @@ export async function isOperator(
 	return parsedResult[0];
 }
 
+/**
+ * Throws an error if the input {@link address} does not support CIS2 format.
+ * @param provider Wallet provider.
+ * @param address Address of a Smart Contract.
+ * @returns undefined.
+ */
 export async function ensureSupportsCis2(
 	provider: WalletApi,
 	address: ContractAddress
-) {
+): Promise<undefined> {
 	const paramsJson = ["CIS-2"];
 	const retValue = await invokeCis2NftContract(
 		provider,
@@ -96,6 +113,14 @@ export async function ensureSupportsCis2(
 	}
 }
 
+/**
+ * Gets token balance of an account.
+ * @param provider Wallet Provider.
+ * @param account Account to check the balance.
+ * @param nftAddress Address of CIS2 smart contract.
+ * @param tokenId Hex encoded Token Id.
+ * @returns Balance of the {@link tokenId} Token for account {@link account} in CIS2 contract {@link nftAddress}
+ */
 export async function balanceOf(
 	provider: WalletApi,
 	account: string,
@@ -123,12 +148,20 @@ export async function balanceOf(
 	return parsedResult[0];
 }
 
+/**
+ * Gets token Metadata Url.
+ * @param provider Wallet Provider.
+ * @param account Account address.
+ * @param nftContractAddress Contract Address.
+ * @param tokenId Hex encoded Token Id.
+ * @returns Token Metadata {@link MetadataUrl}
+ */
 export async function getTokenMetadata(
 	provider: WalletApi,
 	account: string,
 	nftContractAddress: ContractAddress,
 	tokenId: string
-) {
+): Promise<MetadataUrl> {
 	const params = [tokenId];
 	const retValue = await invokeCis2NftContract(
 		provider,
@@ -140,6 +173,16 @@ export async function getTokenMetadata(
 	return new Cis2Deserializer(retValue).readTokenMetadata();
 }
 
+/**
+ * Updates the operator for Account: {@link account}
+ * in Contract: {@link nftContractAddress} to Contract: {@link marketAddress}
+ * @param provider Wallet Provider.
+ * @param account Account address.
+ * @param marketAddress Marketplace contract address.
+ * @param nftContractAddress CIS2-NFT contract address.
+ * @param maxContractExecutionEnergy Max energy to spend on the operation.
+ * @returns Transaction outcomes {@link Record<string, TransactionSummary>}
+ */
 export async function updateOperator(
 	provider: WalletApi,
 	account: string,
@@ -168,10 +211,21 @@ export async function updateOperator(
 		nftContractAddress,
 		MethodName.updateOperator,
 		maxContractExecutionEnergy,
-		toGtu(BigInt(0))
+		BigInt(0)
 	);
 }
 
+/**
+ * Mints an new NFT in Contract: {@link nftContractAddress}
+ * represented by Id: {@link tokenId} and Metadata: {@link tokenMedataUrl}
+ * @param provider Wallet Provider.
+ * @param account Account address.
+ * @param tokenId Token Id
+ * @param tokenMedataUrl Token Metadata Url & Hash
+ * @param nftContractAddress CIS-NFT contract address.
+ * @param maxContractExecutionEnergy Max allowed energy ot Minting.
+ * @returns Transaction outcomes {@link Record<string, TransactionSummary>}
+ */
 export async function mintNft(
 	provider: WalletApi,
 	account: string,
@@ -179,7 +233,7 @@ export async function mintNft(
 	tokenMedataUrl: MetadataUrl,
 	nftContractAddress: ContractAddress,
 	maxContractExecutionEnergy = BigInt(9999)
-) {
+): Promise<Record<string, TransactionSummary>> {
 	const paramJson = {
 		owner: {
 			Account: [account],
@@ -202,17 +256,20 @@ export async function mintNft(
 		nftContractAddress,
 		MethodName.mint,
 		maxContractExecutionEnergy,
-		toGtu(BigInt(0))
+		BigInt(0)
 	);
 }
 
+/**
+ * Invokes a CIS2 Smart Contract.
+ */
 async function invokeCis2NftContract<T>(
 	provider: WalletApi,
 	contract: ContractAddress,
 	methodName: MethodName,
 	params: T,
 	invoker?: ContractAddress | AccountAddress
-) {
+): Promise<Buffer> {
 	return invokeContract(
 		provider,
 		SCHEMA_BUFFER,
@@ -224,6 +281,9 @@ async function invokeCis2NftContract<T>(
 	);
 }
 
+/**
+ * Updates a CIS2 Smart Contract.
+ */
 async function updateCis2NftContract<T>(
 	provider: WalletApi,
 	paramJson: T,
@@ -231,8 +291,8 @@ async function updateCis2NftContract<T>(
 	contractAddress: ContractAddress,
 	methodName: MethodName,
 	maxContractExecutionEnergy: bigint,
-	amount: GtuAmount
-) {
+	ccdAmount: bigint
+): Promise<Record<string, TransactionSummary>> {
 	return updateContract(
 		provider,
 		CONTRACT_NAME,
@@ -242,17 +302,21 @@ async function updateCis2NftContract<T>(
 		contractAddress,
 		methodName,
 		maxContractExecutionEnergy,
-		amount
+		ccdAmount
 	);
 }
 
+/**
+ * Checks if an input hex encoded token id is a valid token id.
+ * @param tokenIdHex Hex encoded token id.
+ * @param size Size of the token in bytes in the Contract.
+ * @returns true if token is valid.
+ */
 export function isValidTokenId(tokenIdHex: string, size = 4): boolean {
 	try {
 		let buff = Buffer.from(tokenIdHex, "hex");
 		let parsedTokenIdHex = Buffer.from(buff.subarray(0, size)).toString("hex");
-		console.log(
-			`input token id; ${tokenIdHex}, parsed token id:${parsedTokenIdHex}`
-		);
+
 		return parsedTokenIdHex === tokenIdHex;
 	} catch (error) {
 		console.error(error);
