+++ market-ui/src/models/MarketplaceDesrializer.ts
@@ -1,6 +1,9 @@
 import { Cis2Deserializer } from "./Cis2Deserializer";
 import { TokenList, TokenListItem } from "./MarketplaceTypes";
 
+/**
+ * Handles Deserialization of Marketplace types from then underlying buffer.
+ */
 export class MarketplaceDeserializer extends Cis2Deserializer {
 	readTokenList(): TokenList {
 		return this.readVector(this.readTokenListItem, 2);
